// src/lib/types.ts
export type MealDataType = {
  id: string;
  title: string;
  creator: string;
  summary: string;
  slug: string;
  image: string;
  instructions: string;
  creator_email: string;
};

export type MealViewType = {
  id: string;
  title: string;
  creator: string;
  summary: string;
  slug: string;
  image: string;
  instructions: string;
  creatorEmail: string;
};